feat(api): 添加用户权限验证并优化数据库操作

+++ src/main/java/top/aiolife/record/pojo/entity/TaskEntity.java
@@ -38,7 +38,7 @@ public class TaskEntity {
     /**
      * 列ID
      */
-    private Integer columnId;
+    private Long columnId;
 
     /**
      * 目标完成时间
